Update patch set 6
Patch Set 6:

has anyone ever found anything useful with bootchart on Android? the counter position seems to be that without all the shell scripts of a traditional unix system, there's not enough going on for bootchart to be worthwhile and we should just remove it and focus on profiling the zygote instead. or just running system-wide perf.

does anyone have counterexamples where bootchart on Android was genuinely useful and (say) perf wouldn't have been more useful?

Patch-set: 6
Label: Verified=0