Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > Using OpenSSL for this would be good, long term. If it's possible,
 > let me know and I might get around to it.

well, you're the crypto guys. you tell me! we just took ownership of the recovery image, so we're definitely keen to remove any duplication we can, especially of this kind of code. the only reason anyone's given for the duplication was that it was to minimize the binary size, but i'd much rather build some subset of openssl than bring in new code (assuming the linker doesn't do a good enough job of throwing away the 99% of openssl we don't need).

Patch-set: 1