Update patch set 1
Patch Set 1:

do you have more info on this, btw? what kernel does "Hardware" disappear in? was this ever in upstream kernels or was this an Android thing? is "Revision" staying in? (and since devices like N9 seem to currently be relying on "Hardware", do we have a plan to ensure we're not relying on it going forward?)

Patch-set: 1