init: add 'ffbm' action trigger when booting in factory mode
'ffbm' mode is intended to for factory test, to allow the system
to boot quickly with minimal services.

Android init process is extended to recognize ffbm boot mode
and start only required services for this mode. init process adds
new action trigger 'ffbm', which will take care of rest of the
initialization process.

Change-Id: If74201f83d663e2f2932720cc3ecb494a61a854e