liblog: prevent message loss by limiting throughput
If too many messages are logged during a short period of time
some of them might be lost. This can result in failures for
applications that read and analyze the log messages.

This patch introduces a process specific environment variable

  LIBLOG_THROUGHPUT_LIMIT=<msg_count>/<delay_millis>

If it is set then the process delays for delay_millis
milliseconds after printing every msg_count number of messages.

On an aosp_x86_64 emulator the following setting should be
enough to set if only one process is actively logging.

  LIBLOG_THROUGHPUT_LIMIT=201/200

This means that the process will print at most 201 * 1/0.200 =
= 1005 messages per second.

Note: the environment variable is chosen over an Android system
property, because the property getter function is implemented in
libcutils that depends on liblog. Using system property in liblog
would result in a circular dependency, which is not good.
Moreover, environment variables are process specific and allow
finer tuning particular processes.

Change-Id: I7cb62a3f4af497215d33964f19fffa2686f58307
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>