Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

This breaks the low overhead API requirements of the logger.

If we are having troubles with systems that are dropping messages, then increase the value of /proc/sys/net/unix/max_dgram_qlen. It is currently set to 300. This setting is by policy since logging at a rate higher than that is considered too chatty and would result in logs that would be exhausted quickly, but the goal is not to drop logs from normal providers. If you find a new but carefully limited value for max_dgram_qlen that prevents dropping of log messages on your platform, please submit a change to system/core/rootdir/init.rc instead. NB: in my initial design, I set this to 10000, got pushback, 300 was from a large body of experiments on Nexus devices, hardly a full set of platforms.

Another option is to write the code to stream into the logs by embedding newlines and merging log messages up to the 4076 payload limit. eg: Do not use 5 separate log lines, where one would do.

See https://android-review.googlesource.com/#/c/125902 for an Abandoned prototype experiment in lower-overhead logging. Rate limiting was one of the considered options for that backend, and this code may have a place there. That code was abandoned because the overhead of a sem_post was in about the same range and not an astronomical improvement as compared to the overhead to send a log message.

The caller always has the option to check the return value of EBUSY (this was true for the kernel and the user space logger), but ill advised because spinning on such would be considered a bad habit ...

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1