Add init support for dm-verity logging mode
Add built-in commands for loading and updating dm-verity state.
Handle dm-verity uevents in ueventd and update the state if we
come across corrupted blocks.

Also update the state if init is about to reboot the device due
to a crashing critical service, because the crash can be caused
by dm-verity if parts of the partition are corrupted.

Needs changes from
  Ibb82953594d234f81ad21c40f524190b88e4ac8f

Change-Id: I5af4918f2f14fdd4d07f51c55837e08111fd3748