Update patch set 1
Patch Set 1:

Yes, it's still used under emulation when there is no GPU emulation (which is still the default by the way). libGLES_android.so is the software-based GLES 1.x implementation, libpixelflinger is the underlying rendering engine, while libGLES_android.so implements the GLES API on top of it.

Patch-set: 1