Update patch set 3
Patch Set 3:

> I agree that seclabel should be optional, not mandatory; whenever
 > possible, we define the transition through policy.  We only need
 > seclabel for executables in the rootfs or for shell commands (and
 > even the latter can be addressed if it is moved into a script with
 > its own file security context).

makes sense. i'm mainly torn on the syntax question. i agree with both sides: reusing the service syntax means less for people to learn, and more consistency; but it also means that information moves out of line and simple stuff looks more complicated than it is. i'm not convinced by the "you'll need the other options" argument, but i do worry that we're already missing the ability to specify supplementary groups, and it is odd that -- if we make the security label optional in the obvious way ("exec user group --label blah" or whatever) -- 'exec' becomes the first init command to take named arguments.

we could use "--", but that seems a bit subtle, and it doesn't let you solve both the supplementary groups _and_ the optional security label: what does "exec a b c d -- command a1 a2" mean, for example? is d a security label or another group?

 > It also seems that this same issue already exists for execonce, not
 > sure when that was added.  Current SELinux policy does not allow
 > init to execute _anything_ without changing domains, so execonce
 > cannot work for executables in the rootfs presently if enforcing.

the plan is to remove execonce as soon as we can move ghartman over to exec.

Patch-set: 3
Label: Verified=0