Update patch set 1
Patch Set 1: Verified+1

i'd like to discuss this change. we're already pulling in code (and would like to pull in more code) that uses liblog and ALOG* for their logging. in recovery mode, we'd like this to go to stdout / stderr etc.

is this the best way to do it ? should we use __libc_format_log or change headers to use use write(stderr/stdout) directly for ALOGE ?

Patch-set: 1
Label: Verified=+1