logd: Leak liblog events to caller even if not selected
If liblog reports an in-band event for data loss, the caller may wish
to interpret the content to help triage data loss. Even if they did
not specifically enable the event log buffer id to be reported back,
we mix it in so that they are aware of the data loss.

Change-Id: I222c1abfefca0181b8b8f430ccd172daf72a6aa1