Test readx/writex (now renamed).
Renamed readx/writex to ReadFdExact/WriteFdExact respectively. These
are similar to android::WriteStringToFd and android::ReadFdToString,
will return an error if the full length cannot be read or written.
Additionally, these operate on C strings.

Move the TEMP_FAILURE_RETRY definition in sysdeps.h out of the
!Windows section. It seems Windows won't actually interrupt a call,
but it's easier to just define it than to #ifdef each call.

Change-Id: Ia8ddffa2a52764a2f9a281c96c937660e002b9b9