Update patch set 7
Patch Set 7:

A separate effort is going on to provide a tool that will act on a bugreport and merge the logs from logd and dmesg by using correction factors that are logged in the kernel messages.

Correlate ‘Suspended for x.xxx seconds’(imprecise), ‘PM: suspend exit YYYY-MM-DD HH:MM:SS.sssssssss UTC’ and ‘PM: suspend entry YYYY-MM-DD HH:MM:SS.sssssssss UTC’messages to reconstruct/correct for the timestamps in dmesg in order to correct the information.

Patch-set: 7