Update patch set 1
Patch Set 1:

If we were to set _FILE_OFFSET_BITS=64 it would just be PRId64. AIUI as long as there isn't an API exported that uses off_t there isn't an ABI issue (since there seem to be a few places in the tree where we do set this, but not all).

enh would be able to say for sure though.

Patch-set: 1
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0