Update patch set 1
Patch Set 1: Code-Review+1

I'm not concerned with leaking the UID to the log. It's already part of the data structure, and I thought it was already included in some versions of the logcat output.

I didn't look that closely at the actual code, so only +1. I don't have any concerns here.

Patch-set: 1