Update patch set 1
Patch Set 1:

Thanks, if UID list is not a concern, I will not worry about adding new vectors.

TL;DR

I only grew concerned as UID has currently a confounded access except to system applications. This change does not leak because it remains secure, but it is an additional vector for us to keep secure.

UID is not part of any incoming or outgoing logging data structures, it is maintained internally for security checking from socket credentials. We do leak some of that information in the log statistics (logcat -S on eng or userdebug, UID is associated with PID(s)) but it is filtered. Our goal was to limit exposure in upcoming changes to the statistics output (as it is really superfluous when compared to package name). The _only_ association between UID and names or PID are behind limited access in source (include/private/android_filesystem.h, for android applications only), tools(ps), kernel (/proc) or databases (/data/system/packages.xml). An application can only discover its own UID, and that is it.

Patch-set: 1