Update patch set 4
Patch Set 4: Code-Review+1 Verified+1

Uncomfortable with logcat adjustment, but it was part of my design specification to make the output more intuitive. Have checked logging history, the pattern UID:<number> messages do NOT exist. If someone decides to generate them independently, it would be interesting ... we could in the future add filtration that checks the message time, TID and PID with the message immediately following and make it more specific.

Patch-set: 4
Label: Code-Review=+1
Label: Verified=+1