Windows clang compatibility.
Rather than being __attribute__((error)), just don't define the
function on Windows.

This makes the failure a little less clear if anyone ever tries to
call it on Windows (because it will be a link error rather than a
compile error), but with clang the diagnostics happen pre-optimization
so dead code elimination can't prevent the error from being emitted.

Change-Id: Ie228cf92fd582f4765ebc42d3e7aed93773dcb3e