Fix file descriptor leakage in adbd
adb_auth_init in adb_auth_client.cpp sets FD_CLOEXEC on the control
socket, which prevents the leakage. However if ro.adb.secure
property is unset (as it is on the emulator), adb_auth_init is not
invoked, which results in the control socked fd leaking into any
process started by the deamon (specifically, any command executed
through adb shell).

This change makes the invocation of adb_auth_init unconditional,
and moves the "secure mode" check into adb_auth_init. This way, the
FD_CLOEXEC is set regardless of the mode.

Change-Id: I73ea84977542ddfc4ac20599593ecf3745ae9108