Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

opening and reading the /data/system/packages.list file has been turned into a subroutine that spawns a work-item in a separate reserved privileged thread with semaphore communications between them.

The value of the uid to name translation helped immensely with the relevance of the output from the statistics formatting; we tried it without and ended up with a bunch of [UID] references in the tables.

Patch-set: 1
Label: Verified=+1