Update patch set 2
Patch Set 2:

Igor, I appreciate the comments. This change has already gone into internal master, and I'm just trying to minimize the diffs between master and AOSP so I can get working on other things. Would it be possible to spin up a separate change that attempts to address these issues, check it in and let it propagate back to master?

Patch-set: 2