liblog: retry log message delivery in case of errors
If too many messages are logged during a short period of time
some of them might be lost. This can result in failures for
applications that read and analyze the log messages.

This patch introduces a process specific environment variable

  LIBLOG_RETRY_DELIVERY=<retry_count>/<delay_millis>

If it is set then the process retries sending the message
the specified number of times with the specified delay doubled
every next try.

On the aosp_x86_64 emulator even the minimal settings
  LIBLOG_RETRY_DELIVERY=1/0
are enough to set if only one process is actively logging.

This patch is another solution of the message loss problem
described in https://android-review.googlesource.com/#/c/131020/:
  liblog: prevent message loss by limiting throughput

Change-Id: I5a35dc365d3c852d03ac8dce52f5be78f8869515
Signed-off-by: Yevgeny Rouban <yevgeny.y.rouban@intel.com>