Update patch set 1
Patch Set 1:

This is a straight library replacement for the inline defined in system/core/include/private/android_filesystem_config.h. It will remain unused until we activate it by scrubbing the common code out, and replace it with the prototype for the function in here.

Look at the bug for the list of CLs that preceded this, which covered all the cases where includes being deprecated in android_filesystem_config.h would break the builds, and a shared or static library reference to libcutils added to everyone that used this inline function. Those changes frame us for the transition, but are innocuous right now.

Patch-set: 1
Label: Verified=0