Update patch set 1
Patch Set 1: Code-Review-1

Spam filter not yet cleared, I am awaiting approval for the series of 4 CLs ... so dependent on those clearing.

But, it is not about spam protection, but also about ensuring a one-stop shop for all things comms.

TL;DR

The comms team actually _appreciates_ this because it ensures that _their_ logs are all preserved in radio log id. If we allowed them to leak because of the radio binary's (as in, not under code review control) behavior, they will come after us with pitchforks wondering why we allowed their data to be split up to other log ids.

A heuristic that has become an expectation and a culture.

Patch-set: 1