rootfs: init start logd in post-fs
The earliest point we can start logd is after /system is mounted.
As insurance, we will also make sure logd is started if a
logd-reinit is requested. This results in logd starting at least
4 processes earlier than it does currently, with a tighter
grouping of threads which means we are taking advantage of a
lighter CPU load at the time.

Change-Id: If4f0bd3a53bb4c47500a54d741ca635d87c0c330