Update patch set 1
Patch Set 1:

(2 comments)

The reason behind this change was that the mediaserver was crashing due to malformed mp3:s (same as mentioned in 72299bf). A typical crash looks like:

  GDB 285(MediaServer):

  (gdb-arm) bt	
  #0 0x4010e05c in __bionic_heap_error (msg=0x4013d7fa "ARGUMENT IS INVALID HEAP ADDRESS", function= 
     <optimized out>, p=0x422c5c20)at bionic/libc/bionic/dlmalloc.c:75
  #1 0x40110e06 in dlfree (mem=<optimized out>) at     
     bionic/libc/bionic/../upstreamdlmalloc/malloc.c:4803
  #2 0x4010bf74 in free (mem=<optimized out>) at bionic/libc/bionic/malloc_debug_common.cpp:227
  #3 0x401d3186 in release (flags=0, this=0x422c5c28) at 
     frameworks/native/libs/utils/SharedBuffer.cpp:106
  #4 android::SharedBuffer::release (this=0x422c5c28, flags=0) at 
     frameworks/native/libs/utils/SharedBuffer.cpp:100
  #5 0x401d3794 in android::String8::~String8 (this=0xbed407b0, __in_chrg=<optimized out>) at 
     frameworks/native/libs/utils/String8.cpp:195
  #6 0x406af4de in android::MP3Extractor::getMetaData (this=<optimized out>) at 
     frameworks/av/media/libstagefright/MP3Extractor.cpp:635 

In String8.cpp the *length function is used to determine how much memory to allocate.

The problem is that with invalid utf16 data (e.g. 0xd910, 0x0010, a surrogate pair with invalid second half), the length function will return a size that is smaller than what the conversion function will use.

String8.cpp

95static char* allocFromUTF16(const char16_t* in, size_t len)
...
99    const ssize_t bytes = utf16_to_utf8_length(in, len);
100    if (bytes < 0) {
101        return getEmptyString();
102    }
103
...
110    char* str = (char*)buf->data();
111    utf16_to_utf8(in, len, str);   <<< len is wrong and we'll overwrite the target buffer.
112    return str;

Are you guys suggesting that we should rather fix this issue in String8.cpp than here?

Patch-set: 1