libsparse: support ordered write with layers
In some cases, some blocks need to be written after all other blocks
are written out, e.g. superblocks for ext4 file-system to reduce the
window of a partially written fs.

To support this in a generic way, layers are introduced. Layers are
ordered, blocks are queued to the current layer unless a different
one is specified. Order can be specified by new APIs:

  sparse_file_add_{data,fill,file,fd}_ordered(.., int order, ..)

where the "order" defines the order the blocks on this layer will
be written.

To keep this backward compatible, all blocks will be queued to layer
with order being 0 by default with existing APIs:

  sparse_file_add_{data,fill,file,fd}(...)

Switching to a high order layer is simply to rewind to the start of
the output and write out the blocks on that layer. To support this,
a new chunk type is introduced:

  #define CHUNK_TYPE_REWIND	0xCAC5

Ideally, it would be to extend CHUNK_TYPE_DONT_CARE to be more like
CHUNK_TYPE_SKIP, and allows absolute/relative offset to skip, however,
a new type here would allow minimum change and still keep the code
clean.

The new CHUNK_TYPE_REWIND will only be written to the sparse image file
if there is more than 1 layer, so existing code using the non-ordered
APIs will not be impacted at all.

Re-sparse will be per-layer based, a new layer will start a new sparse
image despite how small the layer is.  And when writing the resparsed
image out - always start from offset 0.

Change-Id: Ib30c81ad8ff5ce8d2148c8eff513fc4743712554
Signed-off-by: Eric Miao <emiao@nvidia.com>