adb: statically link against libc++
libc++.so is not widely available on the host, so we compile against
one built as part of the platform.  This causes problems for adb, which
is distributed through a number of channels - the sdk, distro packages,
downloaded from the build server, or manually copied.  Instead of
forcing all users to handle libc++.so too, statically link against
libc++.

Change-Id: I51b75258653a23558c8b598802005f6c1166a439