Update patch set 1
Patch Set 1:

> this seems reasonable, but i lost contact with my N9 trying to boot
 > with this change. i'll have to look into what went wrong on
 > Monday...

Crap. I don't seem able to reproduce a deadlock with the change, but then I am not using the N9 init scripts either. Without the change I get a 100% reproducible deadlock. The init setup code is rather complicated. Couple of things that worries me but did not prove to be a problem:

* why is there a call to handle_signal in signal_init?
* the while(!wait_for_one_process) in signal_handler is scary given the return -1 in wait_for_one_process.

Patch-set: 1