Update patch set 3
Patch Set 3:

I updated the patch so that it patches the exact same place in master. When this change hits master, the automerger will realize there's no change and turn this into a no-op.

I didn't add "DO NOT MERGE" though. My understanding is that the build team doesn't recommend it's use for AOSP contributions, and in my experience, it causes more problems than it solves. If it generates a merge conflict (unlikely, since it matches internal master identically) I'm happy to resolve it by hand.

Patch-set: 3