Update patch set 1
Patch Set 1:

This is now a tuneable IMHO. We can tighten this up once we clear some offenders, or loosen it up if we are still too impactful. The goal is on release, spammers cleaned up, we are all happy if the threshold is dialed down to 1/32. I am liking this graduated approach to the problem; maybe we should make this ro.logd.prune.threshold or something?

Patch-set: 1