Update patch set 2
Patch Set 2:

yeah, it looks like that is a vogar command line of 778 bytes/32 arguments, so that would go up to 778+32*2=842 bytes, and that plus the vogar overhead takes us over 1024.

i'm not sure where that 1024 limit is coming from. (i mean, i can easily grep for the line that outputs the error, but i don't understand why it's making that check.)

Patch-set: 2