Update patch set 1
Patch Set 1:

Not fundamentally opposed to just calling setcon("u:r:init:s0"); directly from init.cpp.  Generally we don't like to do that, but init is rather special already wrt SELinux and we don't really need the flexibility of changing the context associated with the init process in order to evolve the policy.  init.rc offers little additional flexibility; sure it is "configuration" but it is just as immutable as the init program since they both live in the boot partition.

Patch-set: 1