Update patch set 1
Patch Set 1:

systemd does a setcon() to switch to the init label in Linux distributions, just using a context obtained from the policy via security_compute_create().  There is already similar code in init.cpp for computing the service contexts.  But for this to work, you'll also need to restorecon() the init binary and give it its own type in policy.

Patch-set: 1