Update patch set 3
Patch Set 3:

(4 comments)

Thanks for the review, I have a benchmark, but I did not _this_ code out as I actually expect it to be roughly the same speed if nothing is set as compared to using the public interface. The goal was to improve speed _if_ we can when _something_ is set.

Your suggestion to get the global sequence made available will address _all_ the performance concerns. This will require bionic changes prior to doing this.

I may have to consider taking this out of the dependency chain for the related change for 'logd: filter...'; for this I will rerun the benchmark for the impact of _two_ unset properties and determine if I consider it negligible from the logd performance perspective.

Patch-set: 3