Update patch set 1
Patch Set 1: Code-Review-1

If messages are getting lost on a platform, the answer is to increase the value of max_dgram_qlen

    write /proc/sys/net/unix/max_dgram_qlen 300

in the init.rc file. If this value was not enough, we need to correct it so all will benefit. We added a mechanism to report when/if a message is lost ("liblog" tag messages in the "events" log buffer) with the sole intent to respond by fixing the platform by carefully selecting an increased value until the loss goes away.

The logger behaves much more efficiently if the log messages roll in multiple lines to the maximum message payload length, rather than issuing many single LOG message lines for each. If you _must_ send messages at a high rate, it is suggested that the source code be altered.

There is only _one_ excuse for lost messages, and that is to ratelimit sources of logspam. logspam is whenever an application sustains a logging rate in excess of 5 messages/second.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1