Update patch set 1
Patch Set 1:

max_dgram_qlen is not good because we have to set it up in advance. Even if it is big, we have to check for lost messages reported and if reported we have to reboot the system and start again. This complicates test automation.
I do not see any disadvantage of this solution. Do you? Even if we set delay_millis to 1 the application will end up in less than 1 message per second after 10 retries. So no DOS will happen.

Patch-set: 1