Update patch set 1
Patch Set 1:

(1 comment)

Please find an optimal value for max_dgram_qlen that works for you and lets make sure it is part of init.rc for all. max_dgram_qlen _should_ be set to a reasonable no-loss value for all platforms (rate-limiting requirements notwithstanding; we likely will resist a _very_ large value for that reason).

max_dgram_qlen only needs to be set before logd starts (or restarts). If you want you can at anytime at runtime set max_dgram_qlen to 20000, then issue a kill to logd (restart logd) with the new value. You also have the option after it has restarted to run 'logcat -b <buffer> -G 16M' for example, to set a temporary _very_ large buffer to receive your logspam(tm). You can also blacklist any UID that interferes with your logging with 'logcat -P "~<UID>". These are all options you can use during a development cycle to grab more attention from the logger.

One of the problem with your alteration is that one application (including a untrusted applications) gets priority over the Android logs when it is a shared resource. It will be possible for an application to mount a DOS attack on the logs, there is no security protection. The protection may come in the form of selinux blocking setenv ...

We are on the warpath(sic) right now on _any_ logspammers, so we are understandably concerned about anyone exceeding this rate of logging on a permanent ad-hoc basis.

On a typical delivered system, the throughput of logd is ~5 syscalls, to exceed the current max_dgram_qlen value means that one is logging 300 messages in a period of 1.5ms on a typical modern platform, or 200,000 messages a second into a buffer that is 256KB in size. As such, anyone that sustains this level of logging will be marked as a logspammer and automatically pruned aggressively within logd to provide fairness to the other users of the logs (logspam protection can be turned off by not assigning ~! to the pruning list 'logcat -P ""')

We want to solve your problem(s).

Patch-set: 1