adb: win32: fix daemon acknowledgement
The daemon failed to startup because main.cpp was changed from calling
WriteFile() to android::base::WriteStringToFd(), the later which calls
write() in the C Runtime which by default has stdout in textmode which
does \n to \r\n translation.

The quick fix is to allow either OK\n or OK\r\n.

Other potential fix ideas:
* Go back to using WriteFile() for Windows.
* Turn off textmode for stdout (not sure if this makes sense, probably
too scary if people redirect adb to a file).
* ???

Change-Id: I322fc9eae5d6abbf63f3d5917b0beb2171b5a15c
Signed-off-by: Spencer Low <CompareAndSwap@gmail.com>