Fix insmod module size
read_file() appends a new line character to the end of the buffer it
returns.  This newline character breaks the loading of signed kernel
modules since the kernel looks for a specific string at the end of the
buffer.  This fix reduces the size of the buffer passed to
init_module() by one character, which is how insmod() functioned prior
to the cpp-ification of init.

Change-Id: Ie9d9a7adcd33b66621726aef20c4b8cc51c08be7