Add RandomAccessFile API.
Includes a number of RAII style wrappers around various types of
files.

FdFile is a simple wrapper around a file descriptor.

MappedFile is a wrapper around a memory mapped file.

NullFile can be used in place of /dev/null. It does not actually map
to /dev/null, but instead performs no actions. As such, this is
portable to Windows (though this API still needs to be ported to
Windows).

StringFile is an in memory "file" that is backed by a std::string.

CopyFile copies the contents of one RandomAccessFile to another.

Change-Id: Ica7fdff663462b252703d273bb12247f2b158300