Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

Added more comments regarding the algorithm so it is clear where its weaknesses are. The weakness mutes the worst-UID in favour of simplicity and switched to the regular algorithm when the threshold is hit. It does not purely preserve 'a day', just does not allow ourselves to bend over backwards and preserve more than a day.

The major gain here is that we will not look at a phone that has an uptime of a month and see a long record of expire counts that take away from the trust one may have in the logs.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1