Update patch set 1
Patch Set 1:

(1 comment)

Yeah, I like this a lot more. LGTM aside from the suggested change.

As for including windows.h and logging.h in the wrong order, that won't ever happen if people follow the style guide, and anyone that gets it wrong will be greeted with a compile time failure, so that solves itself nicely.

Verifying now.

Patch-set: 1