Refactor subprocess creation.
The pieces of this code that actually handled forking, duping fds, and
execing were quite similar, with the only real difference being the
child's fd and the handling of the pty. Move the pty handling before
the fork so we can just pass in the fd and share the rest of the code.

Default bool arguments make for confusing callsites, so replace is_pty
with an enum.

Remove a nested `#if !ADB_HOST` and the associated Windows
specialization of device only code.

Change-Id: I6b53fa8e16ebea2e0789bb3fccc7bd6b56f0205a