adb_test/libbase_test: win32: get some tests working
adb_test:

* Get it building on Win32 by implementing mkstemp().

* Fix adb_utils directory_exists test for Windows. The test actually
fails because directory_exists() is not aware of junctions or symlinks,
but I'm not really sure if that is a bad thing (since these are rare on
Windows to begin with).

* Fix crash during transport tests due to mutex not being initialized.

* io tests fail for various reasons (see adb_io_test.cpp for more info).

libbase_test:

* Fix file tests to use fake /proc/version.

* Run StringPrintf %z test on Windows because it passes because we build
with __USE_MINGW_ANSI_STDIO which implements %z.

* I didn't fixup the logging tests: some logging tests fail because when
abort() is called on Windows, by default it pops up UI asking whether a
crash dump should be sent to Microsoft. This is another whole ball of
wax (is the crash dump the right behavior? should the dump be sent to
Google with its Windows crash reporter instead? should libbase just has
some env var override (that will call _set_abort_behavior()) for testing
purposes?).

Change-Id: Iaa2433e5294ff162e0b2aa9fe6e4ec09a6893f7a
Signed-off-by: Spencer Low <CompareAndSwap@gmail.com>