Update patch set 1
Patch Set 1: Code-Review+2

> I have evaluated that I do not need this feature in init anymore to
 > solve a problem I had. I was proposing to do the following:
 > 
 > class_start main
 > class_start late_start
 > trigger ${logd.syslogd}
 > . . .
 > on property:ro.build.type=eng
 > setprop logd.syslogd logcatd
 > 
 > on logcatd
 > mkdir /data/system/logd 0700 logd logd
 > . . .
 > start syslogd
 > 
 > To ensure that syslogd started really late so that there was no
 > chance it could hold on to /data on an encrypted device during vold
 > machinations. But a design changed removed the need for this
 > feature being added here in init.
 > 
 > So, do we want it anyways? @enh@ added exec and magically a need
 > for it showed up ;-}

what i really want to do is fix this once and for all, so all arguments to all builtins are expanded. but i keep failing to get round to it, so i've been happily +2ing special cases as people have needed them.

(and i feel this is very similar to the 'exec' case --- we had a long history of bullshit random little hacks instead of just doing the right thing once.)

the other thing that bugs me is that that constant is wrong. yes, it's what everyone else is using, but it's wrong a${b}c can obviously be larger than the maximum length of a single property. expand_props should be rewritten to use std::string.

but, yeah, this patch is making things less bad than they were, so fine by me.

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2