Update patch set 1
Patch Set 1:

Are you sure no one uses external filters (eg, awk scripts etc) and relies on the the "uid=1234(name)" or the "comm=" to parse out the name and command fields? It seems like the benefits in human readability might be offset by the loss of machine readability.

Patch-set: 1