Enable/disable verity for partitions with squashfs
The problem: currently adb enable-verity and adb disable-verity are
used to enable/disable verity only for ext4 file systems. The
implementation for enabling/disabling verity relays in reading the
super block (file system specific) to get the file system size and
then seek into the end of the file system to update the verity
metadata. The routine used to get the device size does not work with
squashfs file systems.
Solution: make use of squashfs-utils library to parse the super block
for squashfs file system instances and adquire the right file system
size.

Change-Id: I9cbd8b54d6efbcb9394761a5b7fdef3cab36aada