Support for overlayfs when no-verity
We want to add overlayfs support to enable users to make modifications
to read-only file systems like squashfs after using 'adb remount'

Added new option to fstab entry to request for an overlayfs mountpoint
using a predefined base directory to create upper and work directories
for the mountpoint.

On the boot process, if overlayfs mountpoints are requested, first we
check if verity is disabled and then after all original mountpoints
are processed, overlay mountpoints are mounted.

Change-Id: I7787e394d2107650b37e8b8125e500b2d8721d25