Wipe overlayfs on adb disable-verity
The idea is that when a mount point has verity enabled, such mount
point by definition is read only and its contents should not be
modified in any way (including by using overlayfs). When a developer
requests for ‘adb disable-verity’, it expects that the contents of the
file system are unchanged up to this point in time.
This patch makes sure that changes that happened at any previous
instance of overlayfs are wiped before returning from disabling
verity.

Change-Id: I4a856eda110dd95f50d5444b6105ece0e3ae1e10