Update patch set 1
Patch Set 1:

Concerned about the risk to the others if we switch mLogElementsLock this late in the program. As it stands today, the usage is clear, and all the others that call these helpers have the lock held. I view these few holes plugged and would consider a followup cleanup with a more extensive set of unit tests?

Patch-set: 1
Label: Verified=0