Update patch set 1
Patch Set 1:

> I'm nothing less than disappointed to learn this is the state of
 > the code base, if this is the case (of contributors) not taking the
 > small efforts to do the simple checks, how can we rely in the
 > future on this platform to drive cars?

I'm not the component owner, so I can't speak for them, but perhaps you can describe the simple check that you're suggesting?

 > I'm in total agreement with you on this specific change after the
 > effort you put but I can't believe we let this go on like
 > that....should I change to plus instead?

I'm not sure all the details of how the AOSP handles -/+, I'd advise you to check the project documentation to guide your decision.

 > Or should we change the
 > code to check the status, document the function to return the
 > relevant error and all that? I can only suggest because I can't
 > spend too much time on the, your call!

My thinking is that larger architectural changes like this should be driven by the component owner because they're in the best position to understand their future plans for the codebase. In other words, non-component owners could prototype large changes, but it is risky if, say, the component owners are planning to do a larger rewrite or reorganization, or change to exception handling, RAII classes, etc.

Here is another example: I was thinking about cleaning up some of the logging/tracing code, but then I realized that the component owners might be rewriting all of that anyway, so any work I do (without full knowledge of their future plans) would just be wasted. So I let it be until later. Hope this helps explain where I'm coming from. Thanks.

Patch-set: 1